requirements.txt
altair==5.4.1
biopandas==0.5.1
biopython==1.84
bleach==6.2.0
ipywidgets==7.6.3
jupyterlab==4.2.5
kaleido==0.2.1
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
plotly==5.24.1
py3dmol==2.4.0
rcsbsearchapi==2.0.0
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
streamlit==1.39.0
stmol==0.0.9
py3dmol==2.4.0
ipython-genutils==0.2.0
joblib==1.4.2
pyfamsa==0.5.3